Correções requeridas pelo lint
