+++ src/pages/consultoria.tsx
@@ -1,0 +1,12 @@
+import React from 'react'
+import PageTemplate from '../components/templates/pageTemplate'
+
+const Consultoria: React.FC = () => {
+    return (
+        <PageTemplate>
+            <h1>Consultoria</h1>
+        </PageTemplate>
+    )
+}
+
+export default Consultoria
